created zesty-home page, reverted components
