+++ src/views/IndexView/IndexView.js
@@ -11,14 +11,16 @@ import {
   Benefits,
   Hero,
 } from './components';
-import Homepage from 'components/zesty-models/Homepage';
 
 const IndexView = () => {
   const theme = useTheme();
   return (
     <Box sx={{ overflowX: 'hidden' }}>
       <Main bgcolor={'background.paper'}>
-        <Homepage></Homepage>
+        <Hero />
+        <Container>
+          <Services />
+        </Container>
         <Box
           sx={{
             backgroundImage: `linear-gradient(to bottom, ${alpha(
